project refactoring, seperate enum voucher type
