+++ src/main/java/com/kdt/commandLineApp/voucher/VoucherService.java
@@ -11,7 +11,6 @@ import java.util.UUID;
 @Service
 public class VoucherService {
     private VoucherRepository voucherRepository;
-
     @Autowired
     public VoucherService(VoucherRepository voucherRepository) {
         this.voucherRepository = voucherRepository;
